feat: BMW: 刷新Consul时，添加ModifyIndex参数作为CAS锁,store基类改造 --story=120680830

+++ pkg/bk-monitor-worker/internal/metadata/service/datasource.go
@@ -182,7 +182,7 @@ func (d DataSourceSvc) CreateDataSource(dataName, etcConfig, operator, sourceLab
 		tx.Commit()
 	}
 	// 触发consul刷新
-	err = NewDataSourceSvc(&ds).RefreshOuterConfig(context.Background(), uint64(0), nil)
+	err = NewDataSourceSvc(&ds).RefreshOuterConfig(context.Background(), 0, nil)
 	if err != nil {
 		return nil, err
 	}
